separate Navigation logic into a new component
